+++ Source/WebKit/NetworkProcess/webtransport/NetworkTransportSession.h
@@ -40,9 +40,7 @@
 namespace WebKit {
 
 class NetworkConnectionToWebProcess;
-class NetworkTransportBidirectionalStream;
-class NetworkTransportReceiveStream;
-class NetworkTransportSendStream;
+class NetworkTransportStream;
 
 struct SharedPreferencesForWebProcess;
 struct WebTransportSessionIdentifierType;
@@ -86,9 +84,9 @@ private:
     IPC::Connection* messageSenderConnection() const final;
     uint64_t messageSenderDestinationID() const final;
 
-    HashMap<WebTransportStreamIdentifier, Ref<NetworkTransportBidirectionalStream>> m_bidirectionalStreams;
-    HashMap<WebTransportStreamIdentifier, Ref<NetworkTransportReceiveStream>> m_receiveStreams;
-    HashMap<WebTransportStreamIdentifier, UniqueRef<NetworkTransportSendStream>> m_sendStreams;
+    HashMap<WebTransportStreamIdentifier, Ref<NetworkTransportStream>> m_bidirectionalStreams;
+    HashMap<WebTransportStreamIdentifier, Ref<NetworkTransportStream>> m_receiveStreams;
+    HashMap<WebTransportStreamIdentifier, Ref<NetworkTransportStream>> m_sendStreams;
     WeakPtr<NetworkConnectionToWebProcess> m_connectionToWebProcess;
 
 #if PLATFORM(COCOA)
