+++ Source/WebKit/SourcesCocoa.txt
@@ -33,10 +33,8 @@ NetworkProcess/cocoa/NetworkSessionCocoa.mm
 NetworkProcess/cocoa/NetworkTaskCocoa.mm
 NetworkProcess/cocoa/WKURLSessionTaskDelegate.mm
 
-NetworkProcess/webtransport/cocoa/NetworkTransportBidirectionalStreamCocoa.mm
-NetworkProcess/webtransport/cocoa/NetworkTransportReceiveStreamCocoa.mm
-NetworkProcess/webtransport/cocoa/NetworkTransportSendStreamCocoa.mm
 NetworkProcess/webtransport/cocoa/NetworkTransportSessionCocoa.mm
+NetworkProcess/webtransport/cocoa/NetworkTransportStreamCocoa.mm
 
 NetworkProcess/Classifier/WebResourceLoadStatisticsStore.cpp
